Rework the way we override Docker Compose settings
Before, we created a docker-compose.override.yaml on the fly that sets
the relevant service images to the images produced in the CI run and
then merged development/docker-compose.ci-override.yml into it with yq.

Now, we overwrite the images in docker-compose.ci-override.yml to use
the CI images and then merge the override file directly into
docker-compose.yml with yq.

The observable effects of this change are:
1. The hard-coded CI overrides in the workflow now take precedence over
   docker-compose.ci-override.yml settings.
2. Due to semantic differences between Docker Compose's built-in
   override mechanism and just yq-merging two docker-compose files
   together, certain override values will now overwrite rather than
   extend the original values. In particular, the dns field will now
   contain only the values in docker-compose.ci-override, rather than
   the concatenation of the two lists of values.

The first is just a side effect that shouldn't matter. The second will
allow us to actually remove unwanted DNS addresses in CI.